+++ src/temple_refactor/scripts/vertex.glsl
