+++ src/try.html
@@ -9,7 +9,7 @@
     <!-- <link rel="stylesheet" href="style copy.css" /> -->
     <script
       type="module"
-      src="./temple_mix/app.js"
+      src="./temple_refactor/scripts/index.js"
     ></script>
   </head>
   <body>
